refactor(community): improve change-password-form
