+++ apps/community/src/features/profile/services/remotes.ts
@@ -12,11 +12,6 @@ interface MyProfile {
 
 type MyProfileUpdate = Pick<MyProfile, 'phone' | 'email'>
 
-interface MyPassword {
-  originalPassword: string
-  newPassword: string
-}
-
 function getMyProfile() {
   return http.get<MyProfile>(END_POINT.MY_PROFILE)
 }
@@ -25,18 +20,4 @@ function patchMyProfile(data: MyProfileUpdate) {
   return http.patch<MyProfileUpdate>(END_POINT.EDIT_MY_PROFILE, data)
 }
 
-function patchChangePassword(data: {
-  originalPassword: string
-  newPassword: string
-}) {
-  return http.patch(END_POINT.CHANGE_PASSWORD, data)
-}
-
-export {
-  type MyProfile,
-  type MyProfileUpdate,
-  type MyPassword,
-  getMyProfile,
-  patchMyProfile,
-  patchChangePassword,
-}
+export { type MyProfile, type MyProfileUpdate, getMyProfile, patchMyProfile }
